+++ pixiedust/display/datahandler/pandasDataFrameHandler.py
@@ -28,22 +28,6 @@ class PandasDataFrameDataHandler(BaseDataHandler):
     def getFieldNames(self, expandNested=False):
         return dataFrameMisc.getFieldNames(PandasDataFrameAdapter(self.entity), expandNested)
 
-    def getFieldNamesAndTypes(self, expandNested=True, sorted=False):
-        fieldNames = self.getFieldNames(expandNested)
-        fieldNamesAndTypes = []
-        for fieldName in fieldNames:
-            fieldType = "unknown/unsupported"
-            if self.isNumericField(fieldName):
-                fieldType = "numeric"
-            elif self.isDateField(fieldName):
-                fieldType = "date/time"
-            elif self.isStringField(fieldName):
-                fieldType = "string"
-            fieldNamesAndTypes.append((fieldName, fieldType))
-        if sorted:
-            fieldNamesAndTypes.sort(key=lambda x: x[0])
-        return fieldNamesAndTypes
-
     def isNumericField(self, fieldName):
         for y in self.entity.columns:
             if y == fieldName:
